// src/app/core/api/models/holiday-to-event.ts
/* tslint:disable */
export interface HolidayToEvent {
  codBanco?: number;
  fecha?: string;
  fechaActualiza?: string;
  fechaCreacion?: string;
  id?: number;
  idConceptos?: Array<number>;
  idDepartamento?: number;
  idEntidad?: number;
  idMunicipio?: number;
  nacional?: string;
  nombre?: string;
  observacion?: string;
  repiteAnno?: string;
  usuarioActualiza?: string;
  usuarioCreacion?: string;
}